+++ src/collective/dms/mailcontent/dmsmail.py
@@ -189,7 +189,7 @@ def incrementIncomingMailNumber(incomingmail, event):
                                                           'collective.dms.mailcontent.browser.settings.IDmsMailConfig.'
                                                           'incomingmail_talexpression')
         incomingmail.internal_reference_no = internal_reference_no
-        incomingmail.reindexObject(idxs=('Title', 'internal_reference_number', 'SearchableText'))
+        incomingmail.reindexObject(idxs=('Title', 'internal_reference_number', 'SearchableText', 'sortable_title'))
     registry = getUtility(IRegistry)
     registry['collective.dms.mailcontent.browser.settings.IDmsMailConfig.incomingmail_number'] += 1
 
@@ -279,7 +279,7 @@ def incrementOutgoingMailNumber(outgoingmail, event):
                                                           'collective.dms.mailcontent.browser.settings.IDmsMailConfig.'
                                                           'outgoingmail_talexpression')
         outgoingmail.internal_reference_no = internal_reference_no
-        outgoingmail.reindexObject(idxs=('Title', 'internal_reference_number', 'SearchableText'))
+        outgoingmail.reindexObject(idxs=('Title', 'internal_reference_number', 'SearchableText', 'sortable_title'))
     registry = getUtility(IRegistry)
     registry['collective.dms.mailcontent.browser.settings.IDmsMailConfig.outgoingmail_number'] += 1